Add restricted/unrestricted toggle on projects
# Conflicts:
#	front/pages/w/[wId]/conversation/space/[spaceId]/index.tsx

# Conflicts:
#	front/pages/w/[wId]/conversation/space/[spaceId]/index.tsx
