+++ front/pages/w/[wId]/conversation/space/[spaceId]/index.tsx
@@ -391,6 +391,7 @@ export default function SpaceConversations({
                     : []
                 }
                 initialManagementMode={spaceInfo.managementMode}
+                initialIsRestricted={spaceInfo.isRestricted}
               />
             )}
           </TabsContent>
